Moved FormWizard into separate component, renamed to StepsWizard
